% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_national_ref_file.R
\name{get_national_ref_file}
\alias{get_national_ref_file}
\title{Get National Reference File}
\usage{
get_national_ref_file(filename, col_select = NULL)
}
\arguments{
\item{filename}{A character string giving the file name to read. Any file
extension supplied will be ignored and matching is case-insensitive.}

\item{col_select}{Columns to include in the results. You can use the same
mini-language as \code{dplyr::select()} to refer to the columns by name. Use
\code{c()} to use more than one selection expression. Although this
usage is less common, \code{col_select} also accepts a numeric column index. See
\code{\link[tidyselect:language]{?tidyselect::language}} for full details on the
selection language.}
}
\value{
A \link[tibble:tibble-package]{tibble} of the requested National
Reference File.
}
\description{
Read a National Reference File.
}
\examples{
get_national_ref_file("gpprac")
get_national_ref_file("GPprac.csv", col_select = c("praccode", "postcode"))
get_national_ref_file("specialty_groupings")
}
