% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_est.R
\name{get_pop_est}
\alias{get_pop_est}
\title{Get population estimates}
\usage{
get_pop_est(
  level = c("datazone", "intzone", "hscp", "ca", "hb"),
  version = "latest",
  min_year = NULL,
  max_year = NULL
)
}
\arguments{
\item{level}{one of "datazone", "intzone", "hscp", "ca" or "hb"}

\item{version}{default is "latest"}

\item{min_year, max_year}{(optional) filter years}
}
\value{
the pop data as a tibble
}
\description{
Get population estimates
}
\examples{
get_pop_est("datazone")
get_pop_est("hb", min_year = 1995, max_year = 2020)
}
