% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_datazone.R
\name{get_simd_datazone}
\alias{get_simd_datazone}
\title{Get a SIMD DataZone Lookup}
\usage{
get_simd_datazone(datazone_version = "latest", simd_version = "latest")
}
\arguments{
\item{datazone_version}{Default is "latest", otherwise supply a year
e.g. "2011"}

\item{simd_version}{Default is "latest", otherwise supply a version
e.g. "2020v2"}
}
\value{
a \link[tibble:tibble-package]{tibble} of the SIMD DataZone lookup
}
\description{
Get a SIMD DataZone Lookup
}
\examples{
get_simd_datazone()
}
