% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_postcode.R
\name{get_simd_postcode}
\alias{get_simd_postcode}
\title{Get a SIMD Postcode Lookup}
\usage{
get_simd_postcode(
  postcode_version = "latest",
  simd_version = "latest",
  col_select = NULL,
  as_data_frame = TRUE
)
}
\arguments{
\item{postcode_version}{Default is "latest", otherwise supply a tag
e.g. "2023_2"}

\item{simd_version}{Default is "latest", otherwise supply a version
e.g. "2020v2"}

\item{col_select}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:eval_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}

\item{as_data_frame}{Should the function return a \code{tibble} (default) or
an Arrow \link[arrow]{Table}?}
}
\value{
a \link[tibble:tibble-package]{tibble} of the SIMD Postcode lookup
}
\description{
Get a SIMD Postcode Lookup
}
\examples{
get_simd_postcode()
}
