% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{have_access}
\alias{have_access}
\title{Check access to lookups}
\usage{
have_access()
}
\value{
\verb{[logical(1)]} - \code{TRUE} if you have appropriate access and \code{FALSE} otherwise.
}
\description{
\code{have_access()} provides a convenient way to check that you have the
correct access to the necessary folders.
}
\examples{
\dontrun{
  library(phslookups)
  if (have_access()){
    print("Success!")
  } else {
    print("Maybe next time")
  }
}
}
