% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hscp_locality_variables}
\alias{hscp_locality_variables}
\title{HSCP Locality Lookup Variables}
\format{
\subsection{\code{hscp_locality_variables}}{

A character vector with 15 elements
}
}
\usage{
hscp_locality_variables
}
\description{
A character vector of the variable names for the HSCP Locality Lookup.
}
\keyword{datasets}
