% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Function to access metadata}
\usage{
metadata(data)
}
\arguments{
\item{data}{Dataset imported via e.g. \code{get_spd()}.}
}
\value{
Metadata \code{tibble} associated with dataset.
}
\description{
Function to access metadata
}
\examples{
library(phslookups)

\dontrun{
spd <- get_spd()

metadata(spd)
}

}
