% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phslookups-package.R
\docType{package}
\name{phslookups-package}
\alias{phslookups}
\alias{phslookups-package}
\title{phslookups: PHS Lookups}
\description{
Helper functions for reading the PHS standard lookups.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Public-Health-Scotland/phslookups}
  \item \url{https://Public-Health-Scotland.github.io/phslookups/}
  \item Report bugs at \url{https://github.com/Public-Health-Scotland/phslookups/issues}
}

}
\author{
\strong{Maintainer}: James Hayes \email{james.hayes2@phs.scot} (\href{https://orcid.org/0000-0002-5380-2029}{ORCID})

Authors:
\itemize{
  \item James Hayes \email{james.hayes2@phs.scot} (\href{https://orcid.org/0000-0002-5380-2029}{ORCID})
  \item Nicolaos Christofidis \email{nicolaos.christofidis@phs.scot}
  \item Beata Nowok \email{beata.nowok@phs.scot}
}

Other contributors:
\itemize{
  \item Public Health Scotland [copyright holder]
}

}
\keyword{internal}
