% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simd_datazone_variables}
\alias{simd_datazone_variables}
\title{SIMD DataZone file Variables}
\format{
\subsection{\code{simd_datazone_variables}}{

A character vector with 41 elements
}
}
\usage{
simd_datazone_variables
}
\description{
A character vector of the variable names for the latest SIMD DataZone file.
}
\keyword{datasets}
