% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simd_postcode_variables}
\alias{simd_postcode_variables}
\title{SIMD Postcode file Variables}
\format{
\subsection{\code{hscp_locality_variables}}{

A character vector with 42 elements
}
}
\usage{
simd_postcode_variables
}
\description{
A character vector of the variable names for the latest SIMD Postcode file.
}
\keyword{datasets}
