% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spd_variables}
\alias{spd_variables}
\title{Scottish Postcode Directory Variables}
\format{
\subsection{\code{spd_variables}}{

A character vector with 88 elements
}
}
\usage{
spd_variables
}
\description{
A character vector of the variable names for the latest Scottish Postcode
Directory.
}
\keyword{datasets}
