% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hscp_locality.R
\name{get_hscp_locality}
\alias{get_hscp_locality}
\title{Get the HSCP Locality lookup}
\usage{
get_hscp_locality(version = "latest")
}
\arguments{
\item{version}{Default is "latest", otherwise supply a date e.g. "20230804"}
}
\value{
a \link[tibble:tibble-package]{tibble} of the HSCP localities lookup
}
\description{
Get the HSCP Locality lookup
}
\examples{
get_hscp_locality()
}
