% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_est.R
\name{get_population_estimates}
\alias{get_population_estimates}
\alias{get_hb_pop_est}
\alias{get_ca_pop_est}
\alias{get_hscp_pop_est}
\alias{get_dz_pop_est}
\alias{get_iz_pop_est}
\title{Get population estimates}
\usage{
get_hb_pop_est(
  version = "latest",
  min_year = NULL,
  max_year = NULL,
  age_groups = FALSE,
  pivot_wider = FALSE,
  ...
)

get_ca_pop_est(
  version = "latest",
  min_year = NULL,
  max_year = NULL,
  age_groups = FALSE,
  pivot_wider = FALSE,
  ...
)

get_hscp_pop_est(
  version = "latest",
  min_year = NULL,
  max_year = NULL,
  age_groups = FALSE,
  pivot_wider = FALSE,
  ...
)

get_dz_pop_est(
  version = "latest",
  min_year = NULL,
  max_year = NULL,
  age_groups = FALSE,
  pivot_wider = FALSE,
  ...
)

get_iz_pop_est(
  version = "latest",
  min_year = NULL,
  max_year = NULL,
  age_groups = FALSE,
  pivot_wider = FALSE,
  ...
)
}
\arguments{
\item{version}{The version of the population estimates to use
(default: "latest").}

\item{min_year, max_year}{(optional) The minimum and maximum years to
include in the results.}

\item{age_groups}{Logical, indicating whether to aggregate population
estimates by age groups.
If \code{TRUE}, the \code{phsmethods::create_age_groups} function is used.}

\item{pivot_wider}{Optionally reshape the data into a wider format,
summarising population counts by the specified columns.
Allowed values:
\itemize{
\item \code{FALSE} (default): Do not pivot.
\item \code{TRUE} or \code{"all"}: Pivot by both sex and age/age group.
\item \code{"age"}: Pivot by age/age group only.
\item \code{"age-only"}: Pivot by age/age group and aggregate to remove sex.
\item \code{"sex"}: Pivot by sex only.
\item \code{"sex-only"}: Pivot by sex group and aggregate to remove age/age group
}}

\item{...}{Additional arguments passed to \code{\link[phsmethods:create_age_groups]{phsmethods::create_age_groups()}}.}
}
\value{
A tibble containing the filtered and possibly transformed
population data.
}
\description{
These functions retrieve population estimates for different geographic
levels. They read population data from specified files and filter it
based on the input parameters. The functions also allow for grouping by
age and pivoting the data for wider format.
}
