% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_datazone.R
\name{get_simd_datazone}
\alias{get_simd_datazone}
\title{Get Data Zone-SIMD lookup}
\usage{
get_simd_datazone(simd_version = "latest", col_select = NULL)
}
\arguments{
\item{simd_version}{A string defining a SIMD version. The default value
is "latest". Alternatively you can supply a string defining a specific
version. It should follow pattern "YYYY" or "YYYYv2", e.g. "2020v2".}

\item{col_select}{Columns to include in the results. You can use the same
mini-language as \code{dplyr::select()} to refer to the columns by name. Use
\code{c()} to use more than one selection expression. Although this
usage is less common, \code{col_select} also accepts a numeric column index. See
\code{\link[tidyselect:language]{?tidyselect::language}} for full details on the
selection language.}
}
\value{
a \link[tibble:tibble-package]{tibble} of the SIMD DataZone lookup
}
\description{
Read a Data Zone-Scottish Index of Multiple Deprivation (SIMD) lookup file
from cl-out into a tibble.
}
\examples{
get_simd_datazone()
get_simd_datazone(simd_version = "2016")
get_simd_datazone(
  simd_version = "2016",
  col_select = c("DataZone2011", "simd2016rank")
)
}
