% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_datazone.R
\name{get_simd_datazone}
\alias{get_simd_datazone}
\title{Get a SIMD DataZone Lookup}
\usage{
get_simd_datazone(
  datazone_version = "latest",
  simd_version = "latest",
  col_select = NULL
)
}
\arguments{
\item{datazone_version}{Default is "latest", otherwise supply a year
e.g. "2011"}

\item{simd_version}{Default is "latest", otherwise supply a version
e.g. "2020v2"}

\item{col_select}{Columns to include in the results. You can use the same
mini-language as \code{dplyr::select()} to refer to the columns by name. Use
\code{c()} to use more than one selection expression. Although this
usage is less common, \code{col_select} also accepts a numeric column index. See
\code{\link[tidyselect:language]{?tidyselect::language}} for full details on the
selection language.}
}
\value{
a \link[tibble:tibble-package]{tibble} of the SIMD DataZone lookup
}
\description{
Get a SIMD DataZone Lookup
}
\examples{
get_simd_datazone()
}
