% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simd_postcode.R
\name{get_simd_postcode}
\alias{get_simd_postcode}
\title{Get a SIMD Postcode Lookup}
\usage{
get_simd_postcode(
  postcode_version = "latest",
  simd_version = "latest",
  col_select = NULL
)
}
\arguments{
\item{postcode_version}{Default is "latest", otherwise supply a tag
e.g. "2023_2"}

\item{simd_version}{Default is "latest", otherwise supply a version
e.g. "2020v2"}

\item{col_select}{Columns to include in the results. You can use the same
mini-language as \code{dplyr::select()} to refer to the columns by name. Use
\code{c()} to use more than one selection expression. Although this
usage is less common, \code{col_select} also accepts a numeric column index. See
\code{\link[tidyselect:language]{?tidyselect::language}} for full details on the
selection language.}
}
\value{
a \link[tibble:tibble-package]{tibble} of the SIMD Postcode lookup
}
\description{
Get a SIMD Postcode Lookup
}
\examples{
get_simd_postcode()
get_simd_postcode(postcode_version = "2016_1", simd_version = "2012")

library(dplyr)
get_simd_postcode(
  postcode_version = "2016_1",
  simd_version = "2012",
  col_select = c("pc7", starts_with("simd"))
)
}
