% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spd.R
\name{get_spd}
\alias{get_spd}
\title{Get the Scottish Postcode Directory}
\usage{
get_spd(version = "latest", col_select = NULL)
}
\arguments{
\item{version}{A string defining a version to read in. The default value
is "latest" and the latest SPD file available on cl-out will be loaded.
Alternatively you can supply a tag, e.g. "2023_2", to load a specific file.}

\item{col_select}{Columns to include in the results. You can use the same
mini-language as \code{dplyr::select()} to refer to the columns by name. Use
\code{c()} to use more than one selection expression. Although this
usage is less common, \code{col_select} also accepts a numeric column index. See
\code{\link[tidyselect:language]{?tidyselect::language}} for full details on the
selection language.}
}
\value{
A \link[tibble:tibble-package]{tibble} of the Scottish Postcode
Directory lookup file or its selected columns.
}
\description{
Read a Scottish Postcode Directory (SPD) lookup file into a tibble.
}
\examples{
get_spd()
get_spd(version = "2023_2", col_select = c("pc7", "latitude", "longitude"))
}
