% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spd.R
\name{get_spd}
\alias{get_spd}
\title{Get the Scottish Postcode Directory}
\usage{
get_spd(version = "latest")
}
\arguments{
\item{version}{Default is "latest", otherwise supply a tag e.g. "2023_2"}
}
\value{
a \link[tibble:tibble-package]{tibble} of the Scottish Postcode Directory
}
\description{
Get the Scottish Postcode Directory
}
\examples{
get_spd()
}
