% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{load_metadata}
\alias{load_metadata}
\alias{.metadata}
\title{Metadata}
\format{
An object of class \code{list} of length 1.
}
\usage{
load_metadata()

.metadata
}
\value{
List of metadata.
}
\description{
Metadata
}
\keyword{datasets}
