% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd_metadata.R
\docType{data}
\name{spd_metadata}
\alias{spd_metadata}
\title{Data dictionary for the Scottish Postcode Directory (SPD) file}
\format{
Data frame with three columns
\describe{
\item{variable_name}{Variable name.}
\item{variable_label}{Variable label.}
\item{range}{Range of variable values.}
}
}
\usage{
spd_metadata
}
\description{
Variables with their labels and ranges of values
}
\examples{
spd_metadata
}
\keyword{datasets}
